Split the downloading management into own class
